Gingalabs/gingalabs/requirements.txt
openai==1.1.12
asyncio
httpx
python-dotenv
supabase